// routes/logic/classes/interfaces/CsvTemplateConfigFieldInterfaces.ts
export interface CsvTemplateConfigsFieldInterfaces {
    csvFieldId : number;
    csvFieldTemplateId : number;
    csvFieldnumOrder : number;
    csvFieldShortDesc: string;
    csvFieldRequired: number;
}

export interface CsvTemplateConfigsFieldDataInterface {
    ID: number;
    CSV_TEMPLATE_CONFIG_ID: number;
    NUM_IN_ORDER: number;
    SHORT_DESC: string;
    REQUIRED: number;
}